Added timestamp to document metadata (issue-15)
The Cosmos DB system manages the '_ts' property internally. This means that maintaining a latest flag on documents as well as keeping the '_ts' field correct is not possible.

A new '@timestamp' document attribute has been added that is managed by Nebula DB. It is only set when performing a write operation.