Improved handling of store configuration updates (issue-22)
Store configuration updates have been shifted up to the Document DB service and is only triggered once when the DB access services are first started. This prevents re-configuration happening every time a new store implementation is created by .NET Core DI flows. This will be a significant performance improvement, esspecially for services with lots of stores and for large data stores (due to less triggering of index re-builds).

Processing configuration in this way allowed the config manager logic to be simplified as well as the removal of a troublesome call to `Wait()`. The blocking `Wait()` call is the only of it's kind the Nebula library and may caues issues with integrations with other libraries. The worst case being potential for a dead lock.